+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.4.0
+
+### Added
+
+- `lisp-comment-dwim` now (un)comments text as well as s-expressions
+
+### Changed
+
+- `lisp-comment-dwim` comments sub s-expressions, not just the first
+  s-expression on the current line where the point is
+
 ## 1.3.0
 
 ### Changed